Fix makefile from robot

+++ sick/modules/live/sick_cart_align.h
@@ -1,12 +1,14 @@
 #ifndef _SICK_CART_ALIGN_H_
 #define _SICK_CART_ALIGN_H_
 
+#include <stdint.h>
+
 #define SICK_CART_ALIGN_SUCCESS      0
 #define SICK_CART_ALIGN_FAIL         1
 #define SICK_CART_ALIGN_TIMEOUT      2
 
 typedef struct {
-    int status;
+    uint8_t status;
 } sick_cart_align_t;
 
 typedef void (*sick_cart_align_receive_data_callback)(sick_cart_align_t *result);
